refactor(resources): move assistants resource into resources package
