+++ src/gigachat/resources/__init__.py
@@ -1,4 +1,5 @@
 from gigachat.resources._utils import warn_deprecated_resource_api
+from gigachat.resources.assistants import AssistantsAsyncClient, AssistantsSyncClient
 from gigachat.resources.chat import (
     AsyncChatNamespace,
     ChatNamespace,
@@ -8,6 +9,8 @@ from gigachat.resources.chat import (
 
 __all__ = (
     "AsyncChatNamespace",
+    "AssistantsAsyncClient",
+    "AssistantsSyncClient",
     "ChatNamespace",
     "LegacyChatAsyncResource",
     "LegacyChatSyncResource",
